backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
openai>=1.12.0,<2.0.0
llama-index==0.10.20
faiss-cpu==1.8.0
pypdf==4.1.0
llama-index==0.10.20
llama-index-core>=0.10.20,<0.11
llama-index-vector-stores-faiss==0.1.2
faiss-cpu